requirements.txt
psycopg[binary]==3.2.13
python-dotenv==1.0.1
sphinx==8.2.3
